# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Mini ERP API
  version: 0.1.0
  description: Contract-first MVP skeleton for Mini ERP.
servers:
  - url: /api/v1
...
paths:
...
  /platform/businesses/{business_id}:
    get:
      tags: [Platform]
      summary: Get platform-level tenant record
      operationId: getPlatformBusiness
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      responses:
        '200':
          description: Business record
    patch:
      tags: [Platform]
      summary: Update platform-level tenant status metadata
      operationId: updatePlatformBusiness
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      responses:
        '200':
          description: Updated
...
  /branches/{branch_id}:
    get:
      tags: [Branches]
      summary: Get branch
      operationId: getBranch
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Branch
    patch:
      tags: [Branches]
      summary: Update branch
      operationId: updateBranch
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Updated
...
  /service-orders:
    get:
      tags: [Service Orders]
      summary: List service orders
      operationId: listServiceOrders
      parameters:
        - name: branch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: assigned_user_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service order list
    post:
      tags: [Service Orders]
      summary: Create service order
      operationId: createServiceOrder
      responses:
        '201':
          description: Service order created
  /service-orders/summary:
    get:
      tags: [Service Orders]
      summary: Service order status summary
      operationId: serviceOrderSummary
      parameters:
        - name: branch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: assigned_user_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service order status counts
  /service-orders/mine:
    get:
      tags: [Service Orders]
      summary: List service orders assigned to current user
      operationId: listMyServiceOrders
      parameters:
        - name: branch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assigned service order list
  /service-orders/{order_id}:
    get:
      tags: [Service Orders]
      summary: Get service order
      operationId: getServiceOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service order
  /service-orders/{order_id}/transition:
    post:
      tags: [Service Orders]
      summary: Transition service order status
      operationId: transitionServiceOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service order transitioned
        '409':
          description: Invalid status transition
  /service-orders/{order_id}/assign:
    post:
      tags: [Service Orders]
      summary: Assign service order
      operationId: assignServiceOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Service order assigned
        '403':
          description: Forbidden
        '409':
          description: Service order closed
  /service-orders/{order_id}/assignments:
    get:
      tags: [Service Orders]
      summary: List active service order assignments
      operationId: listServiceOrderAssignments
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active assignments
...
components:
  parameters:
    BusinessId:
      name: business_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BranchId:
      name: branch_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...